requirements.txt
fastapi[all]
uvicorn[all]
bcrypt
motor
aioredis
aiofiles
asyncio
python-dotenv 
python-jose
pyotp
pytz